separate test environment and production environment. add ci file
