+++ .github/workflows/cmakecicd.yml
@@ -1,39 +1,22 @@
-# 這個工作流程是為了在 Ubuntu 平台上運行 CMake 項目。
-name: CMake on Ubuntu
+name: CMake with Tests on Ubuntu
 
-on:
-  pull_request:
-    branches: [ "main" ]
+on: [push, pull_request]
 
 jobs:
-  build:
-    runs-on: ubuntu-latest  # 指定運行平台為 Ubuntu 的最新版本
+  build-and-test:
+    runs-on: ubuntu-latest
 
     steps:
     - uses: actions/checkout@v3
 
-    - name: Install Qt
-      uses: jurplel/install-qt-action@v3
-      with:
-        version: '6.4.3'
-        modules: 'qthttpserver qtwebsockets'
+    - name: Install dependencies
+      run: sudo apt-get install -y <dependencies>
 
     - name: Configure CMake
-      # 配置 CMake 在 'build' 子目錄中。CMAKE_BUILD_TYPE 只在使用單配置生成器（如 make）時需要。
-      run: |
-        cmake -B ${{ github.workspace }}/build \
-        -DCMAKE_BUILD_TYPE=Release \
-        -S ${{ github.workspace }}
+      run: cmake -B ${{ github.workspace }}/build -DCMAKE_BUILD_TYPE=Release
 
     - name: Build
-      # 使用給定配置構建程式。
       run: cmake --build ${{ github.workspace }}/build
 
-    - name: Package
-      run: windeployqt ${{ github.workspace }}/build/Release/ArmyChess.exe
-
-    - name: Upload
-      uses: actions/upload-artifact@v3
-      with:
-          name: my-project-ubuntu
-          path: ${{ github.workspace }}/build/Release
+    - name: Run Tests
+      run: ctest -C Release --verbose
