+++ src/test/SumTest.cpp
@@ -1,11 +1,11 @@
 #include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 
-#include "../main/sum.hpp"
+#include "../main/sum/sum.hpp"
 
 TEST(FIRSTTEST, SUMTEST){
     int a = 10;
-    int b = 20;
+    int b = 15;
 
     ASSERT_EQ(sum(a, b), a + b);
 }
